requirements.txt
#pip3 install torch==1.8.2+cu111 torchvision==0.9.2+cu111 torchaudio==0.8.2 -f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torchvision==0.9.2+cu111
torchaudio==0.8.2

pyyaml
tqdm
timm==0.4.12
thop  # FLOPS computation

tensorboardx
six
tqdm


